+++ commands/utility/purge.js
@@ -1,22 +1,16 @@
 // Requirements
-const discord = require('discord.js');
-const settings = require('../../settings.json');
+const embedSender = require('../../utilities/embedSender.js');
 
 exports.run = async(client, message, args) => {
+    if(args.length != 1) {
+        embedSender.sendMessageToUser(message, 'Purge Messages', 'Please specify the number of messages you wish to remove.\nE.g.: "~purge 42".');
+        return;
+    }
     const amountOfMessages = parseInt(args.join(' '));
-    const finalAmountOfMessage = Math.min(100, amountOfMessages);
-    const moderatorLog = client.channels.find('name', settings.logChannels.moderator);
+    const finalAmountOfMessage = Math.min(100, amountOfMessages + 1);
 
     message.channel.bulkDelete(finalAmountOfMessage).then(messages => {
-        if(moderatorLog) {
-            const embed = new discord.RichEmbed()
-            .setTimestamp()
-            .setColor(settings.messageColors.colorSuccess)
-            .setTitle(`Event: Purge Messages`)
-            .setDescription(`**__Moderator__**: <@${message.author.id}>\n` +
-                `**__Message__**: Successfully purged ${messages.size} messages.`);
-            moderatorLog.send(embed);
-        }
+        embedSender.logModerator(message, 'Purge Messages', `Successfully purged ${messages.size} messages.`);
     }).catch(console.error);
 };
 
@@ -30,7 +24,7 @@ exports.config = {
 
 exports.help = {
     name: 'purge',
-    description: 'Tries to purge x amount of messages from a given channel.',
+    description: 'Tries to purge x amount of messages from the channel this command is posted in.',
     usage: 'purge <amount of messages>\n\n' +
         '<amount of messages>: Amount of messages you want to purge (limit 100).'
 };
